# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Store Directory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SignupBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserResponseRole =
  (typeof UserResponseRole)[keyof typeof UserResponseRole];

export const UserResponseRole = {
  owner: "owner",
  admin: "admin",
} as const;

export interface UserResponse {
  id: number;
  email: string;
  name: string;
  role: UserResponseRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserResponse;
}

export type StoreStatus = (typeof StoreStatus)[keyof typeof StoreStatus];

export const StoreStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Store {
  id: number;
  storeName: string;
  ownerName: string;
  address: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  imageUrl: string | null;
  discountPercentage: number;
  status: StoreStatus;
  /** @nullable */
  rejectionReason: string | null;
  ownerId: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStoreBody {
  storeName: string;
  ownerName: string;
  address: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  discountPercentage: number;
}

export interface UpdateStoreBody {
  storeName?: string;
  ownerName?: string;
  address?: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  discountPercentage?: number;
}

export interface RejectStoreBody {
  reason?: string;
}

export interface AdminStats {
  totalStores: number;
  pendingStores: number;
  approvedStores: number;
  rejectedStores: number;
  averageDiscount: number;
  recentActivity: Store[];
}

export type ListStoresParams = {
  search?: string;
  minDiscount?: number;
  maxDiscount?: number;
};

export type AdminListStoresParams = {
  status?: AdminListStoresStatus;
  search?: string;
};

export type AdminListStoresStatus =
  (typeof AdminListStoresStatus)[keyof typeof AdminListStoresStatus];

export const AdminListStoresStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;